+++ NEWS.md
@@ -1,3 +1,7 @@
+# Current
+
+- An advanced plotting tutorial
+
 # 24.05.0
 
 - Faster discarding of maths that is too long to read