+++ NEWS.md
@@ -5,7 +5,12 @@
 - More menu icons
 - A better dialogue if maxima isn't found
 - Better search for the maxima binary
-- Find Gnuplot on Windows if wxMaxima is packaged alone.
+- Find Gnuplot on Windows if wxMaxima is packaged alone. This
+  is needed for the 'Popout interactively' function of wxdraw().
+- Security: Use a standard function (CreateTempFileName) for
+  creating a temporary file for the 'popout interactively'
+  function of wxdraw.
+- Security: Add a random number to lisp generated temporary files.
 
 # 24.05.0