+++ NEWS.md
@@ -11,7 +11,7 @@
   creating a temporary file for the 'popout interactively'
   function of wxdraw.
 - Security: Add a random number to lisp generated temporary files.
-- Fix the 'Gnuplot command window' (#771, #1680 and #1732).
+- Fix the 'Gnuplot command window' (on Windows) (issues #771, #1680 and #1732).
 
 # 24.05.0